Mar21/azsdkpython/requirements.txt
azure-mgmt-resource
azure-identity
azure-keyvault-secrets